Split exec config into its own file
